Move proposal to internal part
