+++ shared_model/interfaces/iroha_internal/proposal.hpp
